DirectionService Imp Implemented #3

+++ src/main/java/com/example/proyectotesting/service/CategoryService.java
@@ -13,7 +13,7 @@ public interface CategoryService {
     
     Category save(Category category);
 
-    Integer count();
+    Long count();
 
     boolean deleteById(Long id);
 }
